Move comps to better locations
