+++ hn_next/src/app/page.tsx
@@ -1,11 +1,5 @@
-import { NavBar } from "../components/NavBar";
+import StoryListPage from "./[datatype]/page";
 
 export default function Home() {
-  return (
-    <main className="flex min-h-screen flex-col items-center justify-between p-24 max-w-lg ">
-      <div className="border flex-1">
-        <NavBar />
-      </div>
-    </main>
-  );
+  return <StoryListPage />;
 }
