+++ hn_next/src/components/NavBar.tsx
@@ -1,13 +1,17 @@
+import Link from "next/link";
+
 export function NavBar() {
   return (
-    <nav className="flex w-full justify-between items-center space-x-2">
+    <nav className="flex w-full justify-between items-center space-x-2 border">
       <div className="flex items-center">
-        <h1 className="text-2xl font-bold">HN: Offline</h1>
+        <Link href="/">
+          <h1 className="text-2xl font-bold">HN: Offline</h1>
+        </Link>
       </div>
       <div className="flex items-center gap-2">
-        <p>front</p>
-        <p>day</p>
-        <p>week</p>
+        <Link href="/front">front</Link>
+        <Link href="/day">day</Link>
+        <Link href="/week">week</Link>
       </div>
     </nav>
   );
